🚚 Move Task-related UI tests to test package
In order to better scope the function, the tests are now moved to the
module that composes it.
The compose_dependencies.gradle was updated to contain tests
dependencies for all of them.
