+++ config/dependencies/compose_dependencies.gradle
@@ -26,4 +26,19 @@ dependencies {
     implementation Deps.compose.ui
     implementation Deps.compose.material
     implementation Deps.compose.tooling
+
+    androidTestImplementation(Deps.compose.uiTest) {
+        exclude group: "androidx.core", module: "core-ktx"
+        exclude group: "androidx.fragment", module: "fragment"
+        exclude group: "androidx.customview", module: "customview"
+        exclude group: "androidx.activity", module: "activity"
+        exclude group: "androidx.lifecycle", module: "lifecycle-runtime"
+    }
+    androidTestImplementation(Deps.compose.junit4) {
+        exclude group: "androidx.core", module: "core-ktx"
+        exclude group: "androidx.fragment", module: "fragment"
+        exclude group: "androidx.customview", module: "customview"
+        exclude group: "androidx.activity", module: "activity"
+        exclude group: "androidx.lifecycle", module: "lifecycle-runtime"
+    }
 }
